Moved operatingHours to protected
